Ride-Sharing Project
The Ride-Sharing System is a web and mobile-based platform that connects passengers with drivers for convenient and cost-effective transportation. The platform allows users to book rides, track drivers in real-time, and make secure payments. It aims to improve urban mobility by reducing traffic congestion, optimizing travel routes, and offering an alternative to traditional taxi services.